requirements.txt
# Core dependencies for CLIP training with Hugging Face dataset
torch>=2.0.0
torchvision>=0.15.0
open-clip-torch>=2.20.0
datasets>=2.14.0
pandas>=1.5.0
Pillow>=9.0.0
tqdm>=4.65.0

# Optional but recommended
numpy>=1.24.0

